requirements.txt
numpy
matplotlib
scikit-image
scikit-learn
antspyx
pyyaml